requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.4
databricks-sql-connector==4.2.4
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
google-auth==2.48.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
lz4==4.4.5
numpy==2.4.2
oauthlib==3.3.1
openai==2.16.0
openpyxl==3.1.5
orjson==3.11.6
packaging==24.2
pandas==2.3.3
pinecone==8.0.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybreaker==1.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
thrift==0.20.0
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
